VP-1677: Separate OrdersFilter.vue to vue and ts
