+++ client-app/src/pages/account/views/AccountOrders.vue
@@ -59,7 +59,7 @@ import Vue from "vue";
 import Component from "vue-class-component";
 import { namespace } from "vuex-class";
 import { BvTableCtxObject } from "bootstrap-vue";
-import OrderFilter from "@account/components/OrderFilter.vue";
+import OrderFilter from "@account/components/orders-filter/index.vue";
 import { FETCH_ORDERS, SET_ORDERS_LIST_CONFIG } from "@account/store/modules/orders-list/definitions";
 import { OrdersList, OrdersListConfig, OrdersListFilters } from "@account/store/modules/orders-list/types";
 import { CustomerOrder } from "@common/api/api-clients";
